# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ginko API
  version: 0.1.0
  description: Local-first canonical memory API for AI agents. Endpoints are available under /api and /api/v1 during v0.x; prefer /api/v1 for new integrations.
servers:
  - url: http://127.0.0.1:8787
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
  /api/config:
    get:
      summary: Get effective server configuration
      responses:
        '200':
          description: Effective config
  /api/context:
    post:
      summary: Build compact token-budgeted prompt context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextRequest'
      responses:
        '200':
          description: Prompt-ready context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextResponse'
  /api/feedback:
    post:
      summary: Record feedback for a context response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextFeedback'
      responses:
        '200':
          description: Feedback recorded
  /api/suggest:
    post:
      summary: Suggest durable memories/learnings from prompts and model output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestRequest'
      responses:
        '200':
          description: Memory candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestResponse'
  /api/usage:
    get:
      summary: List memories with retrieval usage counters
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: subject
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: scope
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: event_limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Memories with usage counters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemoryUsageRow'
  /api/search:
    post:
      summary: Search raw memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Query'
      responses:
        '200':
          description: Matching memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
  /api/memories:
    get:
      summary: Search memories with query parameters
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: subject
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: scope
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Matching memories
    post:
      summary: Create or update a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Memory'
      responses:
        '200':
          description: Stored memory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
  /api/memories/{id}:
    get:
      summary: Get memory by ID
      parameters:
        - $ref: '#/components/parameters/MemoryID'
      responses:
        '200':
          description: Memory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '404':
          description: Not found
    delete:
      summary: Forget/delete memory
      parameters:
        - $ref: '#/components/parameters/MemoryID'
      responses:
        '200':
          description: Deleted
  /api/supersede/{id}:
    post:
      summary: Supersede an old memory with a newer one
      parameters:
        - $ref: '#/components/parameters/MemoryID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Memory'
      responses:
        '200':
          description: New memory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
  /api/events:
    get:
      summary: List recent audit events
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /api/documents/{id}/suggest:
    post:
      summary: Suggest memory candidates from document chunks
      parameters:
        - $ref: '#/components/parameters/MemoryID'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChunkSuggestRequest'
      responses:
        '200':
          description: Memory candidates with chunk provenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkSuggestResponse'
  /api/chunks/search:
    post:
      summary: Search RAG chunks/evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChunkSearchRequest'
      responses:
        '200':
          description: Matching chunks with document provenance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChunkSearchResult'
  /api/ingest:
    post:
      summary: Ingest a file or directory into RAG documents/chunks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Ingestion run result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
  /api/ingestion-runs:
    get:
      summary: List RAG ingestion runs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: Ingestion runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IngestionRun'
  /api/documents:
    get:
      summary: List ingested RAG documents
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: Required for `/api` and `/api/v1` endpoints when `server.auth_token` or `server.auth_token_env` is configured; non-loopback binds require a token.
  parameters:
    MemoryID:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    ChunkSuggestRequest:
      type: object
      properties:
        subject: { type: string }
        scope: { type: string }
        limit: { type: integer, default: 10 }
        min_confidence: { type: number, default: 0.65 }
        store: { type: boolean, default: false }
    ChunkSuggestResponse:
      type: object
      properties:
        document:
          $ref: '#/components/schemas/Document'
        candidates:
          type: array
          items:
            type: object
            properties:
              memory:
                $ref: '#/components/schemas/Memory'
              reason: { type: string }
              requires_confirmation: { type: boolean }
        stored:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
    ChunkSearchRequest:
      type: object
      properties:
        text: { type: string }
        document_id: { type: string }
        limit: { type: integer, default: 20 }
    ChunkSearchResult:
      type: object
      properties:
        chunk:
          $ref: '#/components/schemas/Chunk'
        document:
          $ref: '#/components/schemas/Document'
        score: { type: number }
    IngestRequest:
      type: object
      required: [path]
      properties:
        path: { type: string }
        recursive: { type: boolean, default: false }
    IngestionRun:
      type: object
      properties:
        id: { type: string }
        source_path: { type: string }
        recursive: { type: boolean }
        parser: { type: string }
        status: { type: string }
        files_seen: { type: integer }
        documents_created: { type: integer }
        chunks_created: { type: integer }
        error: { type: string }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
    IngestResponse:
      type: object
      properties:
        run:
          $ref: '#/components/schemas/IngestionRun'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/Chunk'
        skipped:
          type: array
          items: { type: string }
    Document:
      type: object
      properties:
        id: { type: string }
        ingestion_run_id: { type: string }
        path: { type: string }
        title: { type: string }
        source_kind: { type: string }
        source_ref: { type: string }
        sha256: { type: string }
        created_at: { type: string, format: date-time }
    Chunk:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        ordinal: { type: integer }
        heading_path: { type: string }
        content: { type: string }
        token_count: { type: integer }
        page_from: { type: integer, nullable: true }
        page_to: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    Source:
      type: object
      required: [kind, ref]
      properties:
        kind: { type: string, example: conversation }
        ref: { type: string, example: session/message }
    Memory:
      type: object
      required: [type, subject, content, source, scope, confidence]
      properties:
        id: { type: string, example: mem_01J }
        type:
          type: string
          enum: [preference, fact, decision, task, note, relationship]
        subject: { type: string, example: botmaster }
        content: { type: string, example: User prefers direct technical answers. }
        source:
          $ref: '#/components/schemas/Source'
        scope:
          type: string
          enum: [global, project, session, private]
        confidence: { type: number, minimum: 0, maximum: 1, example: 0.95 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        valid_from: { type: string, format: date-time, nullable: true }
        valid_until: { type: string, format: date-time, nullable: true }
        supersedes_id: { type: string, nullable: true }
        superseded_by: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        embedding_refs:
          type: object
          additionalProperties: { type: string }
    Query:
      type: object
      properties:
        text: { type: string }
        types:
          type: array
          items: { type: string }
        scopes:
          type: array
          items: { type: string }
        subject: { type: string }
        tags:
          type: array
          items: { type: string }
        limit: { type: integer, default: 20 }
    ContextRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        subject: { type: string }
        scopes:
          type: array
          items: { type: string }
        types:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        max_tokens: { type: integer, default: 1200 }
        max_memories: { type: integer, default: 12 }
    ContextResponse:
      type: object
      properties:
        context_id: { type: string }
        context: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
        estimated_tokens: { type: integer }
        budget_tokens: { type: integer }
        truncated: { type: boolean }
    MemoryUsage:
      type: object
      properties:
        memory_id: { type: string }
        context_uses: { type: integer }
        useful_votes: { type: integer }
        useless_votes: { type: integer }
        last_used_at: { type: string, format: date-time, nullable: true }
    MemoryUsageRow:
      type: object
      properties:
        usage:
          $ref: '#/components/schemas/MemoryUsage'
        memory:
          $ref: '#/components/schemas/Memory'
    ContextFeedback:
      type: object
      required: [context_id, useful]
      properties:
        context_id: { type: string }
        useful: { type: boolean }
        memory_ids_used:
          type: array
          items: { type: string }
        source:
          $ref: '#/components/schemas/Source'
    SuggestRequest:
      type: object
      properties:
        subject: { type: string }
        scope: { type: string }
        user_prompt: { type: string }
        assistant_response: { type: string }
        llm_inference: { type: string }
        max_candidates: { type: integer }
        min_confidence: { type: number }
    SuggestResponse:
      type: object
      properties:
        candidates:
          type: array
          items:
            type: object
            properties:
              memory:
                $ref: '#/components/schemas/Memory'
              reason: { type: string }
              requires_confirmation: { type: boolean }
    Event:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        payload: { type: string }
        source:
          $ref: '#/components/schemas/Source'
        created_at: { type: string, format: date-time }